Reorganise code
