+++ .gitignore
@@ -1,5 +1,5 @@
 .DS_Store
-build
+dist
 lerna-debug.log
 node_modules
 npm-debug.log
