+++ infra/manual-demo/index.html
@@ -2,6 +2,6 @@
   <body>
     <h1>Dev Build</h1>
     <p><a href="prod.html">Click Here for Prod</a></p>
-    <script src="../../../build/browser-bundles/workbox-core.dev.js"></script>
+    <script src="../../packages/workbox-core/dist/browser/workbox-core.dev.js"></script>
   </body>
 </html>
