+++ packages/workbox-core/.npmignore
@@ -1,0 +1,3 @@
+_browser.mjs
+test/
+.npmignore
