+++ packages/workbox-core/_browser.mjs
@@ -1,0 +1,7 @@
+import defaultExport from './index.mjs';
+import WorkboxError from './models/WorkboxError.mjs';
+
+export default {
+  default: defaultExport,
+  WorkboxError,
+};
