+++ packages/workbox-core/index.mjs
@@ -1,0 +1,3 @@
+import WorkboxCore from './WorkboxCore.mjs';
+
+export default new WorkboxCore();
