--- packages/workbox-core/src/_browser.mjs
@@ -1,7 +1,0 @@
-import instance from '../index';
-import WorkboxError from '../WorkboxError';
-
-export default {
-  instance,
-  WorkboxError,
-};
