--- packages/workbox-core/src/index.mjs
@@ -1,3 +1,0 @@
-import WorkboxCore from './controllers/WorkboxCore.js';
-
-export default new WorkboxCore();
